ENH: Adding a processer for evolution files
This processes the evolution files and does the coordinate transformations.
It still needs to be interfaced with Paraview properly. Meaning we need to
figure out how to register the time-varying single-dimension components.
It should be X(t), Y(t), Z(t) and we need to hook that into a sphere source
to show how the satellite is moving throughout the simulation.