ENH: Update streamlines view and coloring
This updates the streamlines and calculations surrounding them. It
uses CellDatatoPointData on the slice, which helps with the representations
of all the variables downstream (Tube/Arrows). Everything is now constrained
to the longitude plane orientation.

Additionally, this adds color based on the polarity of the variable. That needs
a separate colormap because we already have the Br variable as a frontend
selectable quantity.

Finally, the arrows are combined with the streamlines, so turning on/off the
streamlines will control both the lines and arrows.