ENH: Make colorbar horizontal
This sets the colorbar to be horizontal. This has to be done
in the change variable routine otherwise it would only occur
for the first variable chosen and not capture the updates.
We also need to iterate over all "displays" here because I wasn't
able to get it working for a single display selection. (I tried
the last display, and the lon_slice directly and neither worked)