MNT: Factor out common streamline code
This moves the creation of streamlines into a helper function to
avoid repeating the same creation logic for lon/lat slices.

The latitude slices still have some issues with the display going
completely blank if you try to interact with it, so don't display
them for now, but still keep the creation logic in for later.