PERF: Move UpdatePipeline calls into the if-else blocks
The UpdatePipeline will trigger getting all of the necessary data,
but if we aren't showing that object anyways we don't need to trigger
that code path. Moving the update call into the Show() part of the
if/else makes sure we only hit that if we are already "showing" that
object.