FEAT: Add institute to the metadata
This enables an institute to be identified in the metadata of the
runs. It is currently keyed off of "project" which each institute
has specified in different ways, but this may need to be re-evaluated
again later if we start getting overlapping project values or multiples
from a single institute.