FIX: Update model satellite handling
This changes the model satellite handling to be less Enlil-specific,
making a new ModelSatellite class that has a get_position function.
Each model can override this to implement the position how they need.
For now, we only have an EnlilSatellite class because we don't know
about Euhforia.

When changing model runs, we now delete the satellite objects from
the old run and create new satellite objects for the new run because
there may be varying numbers of satellites and updates, so easiest
just to start over.