ENH: Automatically detect Euhforia data
This detects the file types for Euhforia and sets that program appropriately rather than
requiring the frontend to pass in the program explicitly. This may need
to be more robust in the future, but works as a quick fix for now.

Additionally, I believe the Euhforia data has Earth at +1 x-direction
whereas Enlil has Earth at -1, so set that explicitly. We need to
follow up with all data providers to ensure these are correct.