FIX: Remove solar images, fix satellite fieldlines
Remove the solar image code that is no longer used to simplify
what objects need to be kept around.

Satellites were being removed from the view, but the fieldlines
associated with those satellites were not. This updates the code
to remove the fieldlines associated with a satellite when changing
runs.